TASK: Move command `workspace:rebaseOutdated` to `Neos\Neos\Command\WorkspaceCommandController`
